Move simplify_message into a separate file
